+++ service/src/test/org/apache/hive/service/cli/operation/TestOperationLogManager.java
@@ -26,7 +26,7 @@ import java.util.HashMap;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.apache.hive.common.IPStackUtils;
+import org.apache.hadoop.hive.common.IPStackUtils;
 import org.junit.Before;
 import org.junit.Test;
