+++ standalone-metastore/metastore-common/src/test/java/org/apache/hadoop/hive/common/IPStackUtilsTest.java
@@ -16,11 +16,10 @@
  * limitations under the License.
  */
 
-package org.apache.hive.common;
+package org.apache.hadoop.hive.common;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class IPStackUtilsTest {
 
@@ -30,7 +29,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
 
     String loopback = IPStackUtils.resolveLoopbackAddress();
-    assertEquals(IPStackUtils.LOOPBACK_ADDRESS_IPV4, loopback);
+    Assertions.assertEquals(IPStackUtils.LOOPBACK_ADDRESS_IPV4, loopback);
   }
 
   @Test
@@ -39,7 +38,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(true);
 
     String loopback = IPStackUtils.resolveLoopbackAddress();
-    assertEquals(IPStackUtils.LOOPBACK_ADDRESSES_IPV6.get(0), loopback);
+    Assertions.assertEquals(IPStackUtils.LOOPBACK_ADDRESSES_IPV6.get(0), loopback);
   }
 
   @Test
@@ -48,7 +47,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
 
     String loopback = IPStackUtils.resolveLoopbackAddress();
-    assertEquals(IPStackUtils.LOOPBACK_ADDRESS_IPV4, loopback);
+    Assertions.assertEquals(IPStackUtils.LOOPBACK_ADDRESS_IPV4, loopback);
   }
 
   @Test
@@ -57,7 +56,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
 
     String wildcard = IPStackUtils.resolveWildcardAddress();
-    assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, wildcard);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, wildcard);
   }
 
   @Test
@@ -66,7 +65,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(true);
 
     String wildcard = IPStackUtils.resolveWildcardAddress();
-    assertEquals(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0), wildcard);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0), wildcard);
   }
 
   @Test
@@ -75,15 +74,15 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
     
     String wildcard = IPStackUtils.resolveWildcardAddress();
-    assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, wildcard);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, wildcard);
   }
 
   @Test
   void testConcatHostPort() {
-    assertEquals("192.168.1.1:8080", IPStackUtils.concatHostPort("192.168.1.1", 8080));
-    assertEquals("[2001:db8::1]:8080", IPStackUtils.concatHostPort("2001:db8::1", 8080));
-    assertEquals("[::1]:9090", IPStackUtils.concatHostPort("::1", 9090));
-    assertEquals("example.com:443", IPStackUtils.concatHostPort("example.com", 443));
+    Assertions.assertEquals("192.168.1.1:8080", IPStackUtils.concatHostPort("192.168.1.1", 8080));
+    Assertions.assertEquals("[2001:db8::1]:8080", IPStackUtils.concatHostPort("2001:db8::1", 8080));
+    Assertions.assertEquals("[::1]:9090", IPStackUtils.concatHostPort("::1", 9090));
+    Assertions.assertEquals("example.com:443", IPStackUtils.concatHostPort("example.com", 443));
   }
   
   @Test
@@ -92,7 +91,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
 
     String result = IPStackUtils.adaptWildcardAddress(IPStackUtils.WILDCARD_ADDRESS_IPV4);
-    assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
   }
 
   @Test
@@ -101,7 +100,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
 
     String result = IPStackUtils.adaptWildcardAddress(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0));
-    assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
   }
 
 
@@ -111,7 +110,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(true);
 
     String result = IPStackUtils.adaptWildcardAddress(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0));
-    assertEquals(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0), result);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0), result);
   }
 
   @Test
@@ -120,7 +119,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(true);
 
     String result = IPStackUtils.adaptWildcardAddress(IPStackUtils.WILDCARD_ADDRESS_IPV4);
-    assertEquals(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0), result);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0), result);
   }
 
   @Test
@@ -129,7 +128,7 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
 
     String result = IPStackUtils.adaptWildcardAddress(IPStackUtils.WILDCARD_ADDRESS_IPV4);
-    assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
   }
 
   @Test
@@ -138,14 +137,14 @@ class IPStackUtilsTest {
     IPStackUtils.setPreferIPv6Addresses(false);
 
     String result = IPStackUtils.adaptWildcardAddress(IPStackUtils.WILDCARD_ADDRESSES_IPV6.get(0));
-    assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
+    Assertions.assertEquals(IPStackUtils.WILDCARD_ADDRESS_IPV4, result);
   }
 
   @Test
   void testAdaptWildcardAddress() {
-    assertEquals("192.168.1.1", IPStackUtils.adaptWildcardAddress("192.168.1.1"));
-    assertEquals("2001:db8::1", IPStackUtils.adaptWildcardAddress("2001:db8::1"));
-    assertEquals("example.com", IPStackUtils.adaptWildcardAddress("example.com"));
+    Assertions.assertEquals("192.168.1.1", IPStackUtils.adaptWildcardAddress("192.168.1.1"));
+    Assertions.assertEquals("2001:db8::1", IPStackUtils.adaptWildcardAddress("2001:db8::1"));
+    Assertions.assertEquals("example.com", IPStackUtils.adaptWildcardAddress("example.com"));
   }
 
     // Test cases for getHostAndPort method
@@ -153,38 +152,38 @@ class IPStackUtilsTest {
   @Test
   void testGetHostAndPortWithIPv4() {
     IPStackUtils.HostPort result = IPStackUtils.getHostAndPort("192.168.1.1:8080");
-    assertEquals("192.168.1.1", result.getHostname());
-    assertEquals(8080, result.getPort());
+    Assertions.assertEquals("192.168.1.1", result.getHostname());
+    Assertions.assertEquals(8080, result.getPort());
   }
 
   @Test
   void testGetHostAndPortWithValidIPv6WithSquaredBrackets() {
     IPStackUtils.HostPort result = IPStackUtils.getHostAndPort("[2001:0db8::1]:8080");
-    assertEquals("2001:0db8::1", result.getHostname());
-    assertEquals(8080, result.getPort());
+    Assertions.assertEquals("2001:0db8::1", result.getHostname());
+    Assertions.assertEquals(8080, result.getPort());
   }
 
   @Test
   void testGetHostAndPortWithValidIPv6WithoutSquaredBrackets() {
     IPStackUtils.HostPort result = IPStackUtils.getHostAndPort("2001:0db8::1:8080");
-    assertEquals("2001:0db8::1", result.getHostname());
-    assertEquals(8080, result.getPort());
+    Assertions.assertEquals("2001:0db8::1", result.getHostname());
+    Assertions.assertEquals(8080, result.getPort());
   }
 
   @Test
   void testGetHostAndPortWithHostname() {
     IPStackUtils.HostPort result = IPStackUtils.getHostAndPort("example.com:80");
-    assertEquals("example.com", result.getHostname());
-    assertEquals(80, result.getPort());
+    Assertions.assertEquals("example.com", result.getHostname());
+    Assertions.assertEquals(80, result.getPort());
   }
 
   @Test
   void testGetHostPortWithInvalidAndPort() {
-    assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getHostAndPort("192.168.1.1:70000"),
+    Assertions.assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getHostAndPort("192.168.1.1:70000"),
         "Port number out of range (0-65535).");
-    assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getHostAndPort("192.168.1.1"),
+    Assertions.assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getHostAndPort("192.168.1.1"),
         "Input does not contain a port.");
-    assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getHostAndPort(":8080"),
+    Assertions.assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getHostAndPort(":8080"),
         "Host address is null or empty.");
   }
 
@@ -192,18 +191,18 @@ class IPStackUtilsTest {
 
   @Test
   void testGetPort() {
-    assertEquals(8080, IPStackUtils.getPort("8080"));
-    assertEquals(65535, IPStackUtils.getPort("65535"));
-    assertEquals(0, IPStackUtils.getPort("0"));
+    Assertions.assertEquals(8080, IPStackUtils.getPort("8080"));
+    Assertions.assertEquals(65535, IPStackUtils.getPort("65535"));
+    Assertions.assertEquals(0, IPStackUtils.getPort("0"));
   }
 
   @Test
   void testGetPortWithInvalidPort() {
-    assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getPort("70000"),
+    Assertions.assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getPort("70000"),
         "Port number out of range (0-65535).");
-    assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getPort("-1"),
+    Assertions.assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getPort("-1"),
         "Port number out of range (0-65535).");
-    assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getPort("abc"),
+    Assertions.assertThrows(IllegalArgumentException.class, () -> IPStackUtils.getPort("abc"),
         "For input string: \"abc\"");
   }
 }
